SERVER-57726 Fix bson_util_test with ASAN builder: left shift of negative value -1

+++ src/mongo/bson/util/simple8b_type_util.cpp
@@ -32,7 +32,7 @@
 namespace mongo {
 
 uint64_t Simple8bTypeUtil::encodeInt64(int64_t val) {
-    return (val << 1) ^ (val >> 63);
+    return (uint64_t(val) << 1) ^ (val >> 63);
 }
 
 int64_t Simple8bTypeUtil::decodeInt64(uint64_t val) {
